+++ packages/i18n/I18nDecorator/I18nDecorator.js
@@ -1,40 +1,29 @@
 /**
- * Provides a HOC interface for internationalization.
+ * Adds Internationalization (I18N) support to an application using ilib.
  *
  * @module i18n/I18nDecorator
- * @exports contextTypes
  * @exports I18nDecorator
+ * @exports I18nContextDecorator
+ * @exports I18nContext
  */
 
-import hoc from '@enact/core/hoc';
 import {on, off} from '@enact/core/dispatcher';
-import React from 'react';
-import PropTypes from 'prop-types';
+import hoc from '@enact/core/hoc';
 import {Publisher, contextTypes as stateContextTypes} from '@enact/core/internal/PubSub';
+import PropTypes from 'prop-types';
+import React from 'react';
 
 import ilib from '../src/index.js';
 import {isRtlLocale, updateLocale} from '../locale';
 
 import getI18nClasses from './getI18nClasses';
 
-/**
- * An object that exports the default context validation rules.
- *
- * These must be applied to any child components that wish to receive the i18n context.
- *
- * ```
- * import {contextTypes} from '@enact/i18n/I18nDecorator';
- * ...
- * myComponent.contextTypes = contextTypes;
- * ```
- *
- * @memberof i18n/I18nDecorator
- * @public
- */
 const contextTypes = {
 	rtl: PropTypes.bool,
 	updateLocale: PropTypes.func
 };
+
+const I18nContext = React.createContext(null);
 
 /**
  * A higher-order component that is used to wrap the root element in an app. It provides an `rtl` member on the
@@ -48,8 +37,10 @@ const contextTypes = {
  * @hoc
  * @public
  */
-const IntlHoc = hoc((config, Wrapped) => {
-	return class I18nDecorator extends React.Component {
+const I18nDecorator = hoc((config, Wrapped) => {
+	return class extends React.Component {
+		static displayName = 'I18nDecorator'
+
 		static propTypes = /** @lends i18n/I18nDecorator.I18nDecorator.prototype */ {
 			/**
 			 * Classname for a root app element.
@@ -79,7 +70,9 @@ const IntlHoc = hoc((config, Wrapped) => {
 			const locale = props.locale && props.locale !== ilibLocale ? updateLocale(props.locale) : ilibLocale;
 
 			this.state = {
-				locale: locale
+				locale: locale,
+				rtl: isRtlLocale(),
+				updateLocale: this.updateLocale
 			};
 		}
 
@@ -132,11 +125,13 @@ const IntlHoc = hoc((config, Wrapped) => {
 		 */
 		updateLocale = (newLocale) => {
 			const locale = updateLocale(newLocale);
-			this.setState({locale});
-			this.publisher.publish({
+			const updated = {
 				locale,
 				rtl: isRtlLocale()
-			});
+			};
+
+			this.setState(updated);
+			this.publisher.publish(updated);
 		}
 
 		render () {
@@ -149,14 +144,59 @@ const IntlHoc = hoc((config, Wrapped) => {
 			delete props.locale;
 
 			return (
-				<Wrapped {...props} className={classes} />
+				<I18nContext.Provider value={this.state}>
+					<Wrapped {...props} className={classes} />
+				</I18nContext.Provider>
 			);
 		}
 	};
 });
 
-export default IntlHoc;
+const defaultConfig = {
+	localeProp: null,
+	rtlProp: null,
+	updateLocaleProp: null
+};
+
+const I18nContextDecorator = hoc(defaultConfig, (config, Wrapped) => {
+	const {localeProp, rtlProp, updateLocaleProp} = config;
+
+	// eslint-disable-next-line no-shadow
+	return function I18nContextDecorator (props) {
+		return (
+			<I18nContext.Consumer>
+				{(i18nContext) => {
+
+					if (i18nContext) {
+						const {locale, rtl, updateLocale: update} = i18nContext;
+
+						props = Object.assign({}, props);
+						if (localeProp) {
+							props[localeProp] = locale;
+						}
+
+						if (rtlProp) {
+							props[rtlProp] = rtl;
+						}
+
+						if (updateLocaleProp) {
+							props[updateLocaleProp] = update;
+						}
+					}
+
+					return (
+						<Wrapped {...props} />
+					);
+				}}
+			</I18nContext.Consumer>
+		);
+	};
+});
+
+export default I18nDecorator;
 export {
-	IntlHoc as I18nDecorator,
-	contextTypes
+	contextTypes,
+	I18nContext,
+	I18nContextDecorator,
+	I18nDecorator
 };